pyproject.toml
[tool.poetry]
name = "autogen-podcast"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyautogen = "^0.1.14"
python-dotenv = "^1.0.0"
argparse = "^1.4.0"
openai = "^0.28.1"
tiktoken = "^0.5.1"
nltk = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "autogen_podcast.main:main"
tts = "autogen_podcast.modules.tts"

